Move standard and uniform benchmarks to top-level crate
